# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Credity API
  version: 1.1.0
  description: |
    Credity standards-aligned API surface for issuance, presentation, verification,
    status, and asynchronous anchoring.
  x-contract-freeze-date: 2026-02-12
  x-deprecation-policy:
    notice_period_days: 90
    sunset_header_required: true
    communication_channels:
      - changelog
      - dashboard-banner
      - email
servers:
  - url: https://api.credity.in
  - url: http://localhost:5001
    description: Issuer local service
  - url: http://localhost:5003
    description: Recruiter local service
...
paths:
...
  /v1/verifications/{id}:
    get:
      tags: [verification]
      summary: Get verification result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification record
        "404":
          description: Not found
  /v1/status/bitstring/{listId}:
    get:
      tags: [status]
      summary: Get bitstring status list
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Status list
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...